Moved ProductList to Components as it is no longer bespoke
